feat(form): submit medication data to the database
