+++ .circleci/config.yml
@@ -25,7 +25,7 @@ jobs:
           command: npm test
       - run:
           name: Build
-          command: npm run build
+          command: CI=false npm run build
       # - run:
       #       name: Firebase Deploy
       #       command: ./node_modules/.bin/firebase deploy --token "$FIREBASE_TOKEN" --only hosting
