+++ src/utils/helper.js
@@ -1,28 +1,28 @@
 /** @format */
 
 // let objData = {
-//   oftenFreq: "12",
-//   name: "aspirin",
-//   type: "tablet",
-//   strength: "10",
-//   unit: "mg",
-//   description: "white round tablet",
-//   amount: "1",
-//   ampm1: "AM",
-//   hour1: "08",
-//   minute1: "00",
-//   ampm2: "PM",
-//   hour2: "09",
-//   minute2: "30",
-//   ampm11: "PM",
-//   hour11: "09",
-//   minute11: "30",
-//   ampm15: "PM",
-//   hour15: "09",
-//   minute15: "30",
-//   priority: "Medium",
-//   notes: "Eat with food.",
-//   supply: "10",
+// 	oftenFreq: "12",
+// 	name: "aspirin",
+// 	type: "tablet",
+// 	strength: "10",
+// 	unit: "mg",
+// 	description: "white round tablet",
+// 	amount: "1",
+// 	ampm1: "AM",
+// 	hour1: "08",
+// 	minute1: "00",
+// 	ampm2: "PM",
+// 	hour2: "09",
+// 	minute2: "30",
+// 	ampm11: "PM",
+// 	hour11: "09",
+// 	minute11: "30",
+// 	ampm15: "PM",
+// 	hour15: "09",
+// 	minute15: "30",
+// 	priority: "Medium",
+// 	notes: "Eat with food.",
+// 	supply: "10",
 // };
 function convertData(objData) {
 	for (const prop in objData) {
